Added transfer api for transfering moneys from one account to another with currency conversion

+++ src/main/java/com/krotos/MoneyTransfer/exchangeRates/ExchangeRates.java
@@ -2,6 +2,7 @@ package com.krotos.MoneyTransfer.exchangeRates;
 
 import com.krotos.MoneyTransfer.Currency;
 
+import java.math.BigDecimal;
 import java.util.Iterator;
 import java.util.List;
 
@@ -12,15 +13,16 @@ class ExchangeRates {
         this.rates = rates;
     }
 
-    double getUSDValue(Currency currency){
+    BigDecimal getUSDValue(Currency currency){
         return rates.stream().filter(r->r.getCurrency()==currency).findFirst().get().getUSDValue();
     }
 
     void updateRate(Currency currency, double usdValue){
-        rates.stream().filter(r->r.getCurrency()==currency).findFirst().get().setUSDValue(usdValue);
+        rates.stream().filter(r->r.getCurrency()==currency).findFirst().get().setUSDValue(BigDecimal.valueOf(usdValue));
     }
 
     Iterator<Rate> getIterator(){
         return rates.iterator();
     }
+
 }
